Defer and delegate event listeners
Limit the amount of `<trix-editor>` element adapting necessary by
declaring event handlers on the `TrixMentionsElement` directly. For
event listeners that were previously attached to `this.input`, defer
checks that the bubbling events' targets are references to `this.input`
until bubble-time.

Similarly, when receiving a `fragment:` option in the
`trix-mentions-change` event's `provide` Promise, check whether or not
it [isConnected][]. During combobox activation, If it's already
connected to the document, remove its `[hidden]` attribute. If it's a
fragment that isn't yet connected, append it to the expander. During
deactivation, set `[hidden]` when the menu was already connected, or
remove it if it started out as a fragment.

[isConnected]: https://developer.mozilla.org/en-US/docs/Web/API/Node/isConnected