Mention `window.Trix` in README.md
Currently, the Trix package's custom element definition is not
idempotent, so calling `import "trix"` multiple times results in errors.

Due to that issue coupled with the package's current lack of TypeScript
information, our implementation accesses the `Trix` dependency through
`window.Trix`.

To guarantee that access, consumers should follow the instructions in
the README to install that class onto the global `window` object.