Bugfix: Read `[data-trix-attachment]` from commit
The `getJSONAttribute` helper's return type was `any`, which includes
`undefined`. Unfortunately, that return type obscured the fact that the
initial implementation lacked a `return` statement in the `try { ... }`
block, resulting in an `undefined` value when it should have otherwise
been the contents of a committing element's `[data-trix-attachment]`
attribute.

To resolve the underlying issue, replace the `any` return type with a
`Record<string, unknown>`, which guarantees that _something_ is
returned.