Position listbox when match key is first detected
Prior to this commit, the initial state of the expanded
`[role="listbox"]` element was not navigable due to an attempt to access
a cursor position outside the `Trix.Editor` instance's selection range.

This commit positions the cursor within that range by deducting 1
position from the maximum edge of the range.